extract facets into utils/facets.ts
